+++ lib/oban/migrations.ex
@@ -6,34 +6,47 @@ defmodule Oban.Migrations do
   @initial_version 1
   @current_version 2
 
-  def up(version \\ @current_version) do
-    change(get_current()..version, :up)
+  @default_opts %{prefix: "public", version: @current_version}
+
+  def up(opts \\ []) when is_list(opts) do
+    opts
+    |> merge_defaults()
+    |> change(:up)
   end
 
-  def down(version \\ @current_version) do
-    change(get_current()..version, :down)
+  def down(opts \\ []) when is_list(opts) do
+    opts
+    |> merge_defaults()
+    |> change(:down)
   end
 
-  defp change(range, direction) do
-    for index <- range do
-      [__MODULE__, "V#{index}"]
-      |> Module.safe_concat()
-      |> apply(direction, [])
-    end
+  defp merge_defaults(opts) do
+    opts = Map.merge(@default_opts, Map.new(opts))
+
+    Map.put(opts, :range, get_current(opts.prefix)..opts.version)
   end
 
-  defp get_current do
+  defp get_current(prefix) do
     result =
       repo().query("""
       SELECT description
       FROM pg_class
       LEFT JOIN pg_description ON pg_description.objoid = pg_class.oid
-      WHERE relname = 'oban_jobs'
+      LEFT JOIN pg_namespace ON pg_namespace.oid = pg_class.relnamespace
+      WHERE pg_class.relname = 'oban_jobs' AND pg_namespace.nspname = '#{prefix}'
       """)
 
     case result do
       {:ok, %{rows: [[version]]}} when is_binary(version) -> String.to_integer(version)
       _ -> @initial_version
+    end
+  end
+
+  defp change(%{prefix: prefix, range: range}, direction) do
+    for index <- range do
+      [__MODULE__, "V#{index}"]
+      |> Module.safe_concat()
+      |> apply(direction, [prefix])
     end
   end
 
@@ -48,12 +61,16 @@ defmodule Oban.Migrations do
       end
     end
 
-    def up do
+    def up(prefix) do
+      execute "CREATE SCHEMA IF NOT EXISTS #{prefix}"
+
       execute """
       DO $$
       BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_type WHERE typname = 'oban_job_state') THEN
-          CREATE TYPE oban_job_state AS ENUM (
+      IF NOT EXISTS (SELECT 1 FROM pg_type
+                     WHERE typname = 'oban_job_state'
+                       AND typnamespace = '#{prefix}'::regnamespace::oid) THEN
+          CREATE TYPE #{prefix}.oban_job_state AS ENUM (
             'available',
             'scheduled',
             'executing',
@@ -65,9 +82,9 @@ defmodule Oban.Migrations do
       END$$;
       """
 
-      create_if_not_exists table(:oban_jobs, primary_key: false) do
+      create_if_not_exists table(:oban_jobs, primary_key: false, prefix: prefix) do
         add :id, :bigserial, primary_key: true
-        add :state, :oban_job_state, null: false, default: "available"
+        add :state, :"#{prefix}.oban_job_state", null: false, default: "available"
         add :queue, :text, null: false, default: "default"
         add :worker, :text, null: false
         add :args, :map, null: false
@@ -81,18 +98,18 @@ defmodule Oban.Migrations do
         add :completed_at, :utc_datetime_usec
       end
 
-      create_if_not_exists index(:oban_jobs, [:queue])
-      create_if_not_exists index(:oban_jobs, [:state])
-      create_if_not_exists index(:oban_jobs, [:scheduled_at])
+      create_if_not_exists index(:oban_jobs, [:queue], prefix: prefix)
+      create_if_not_exists index(:oban_jobs, [:state], prefix: prefix)
+      create_if_not_exists index(:oban_jobs, [:scheduled_at], prefix: prefix)
 
       execute """
-      CREATE OR REPLACE FUNCTION oban_jobs_notify() RETURNS trigger AS $$
+      CREATE OR REPLACE FUNCTION #{prefix}.oban_jobs_notify() RETURNS trigger AS $$
       DECLARE
         channel text;
         notice json;
       BEGIN
         IF (TG_OP = 'INSERT') THEN
-          channel = 'oban_insert';
+          channel = '#{prefix}.oban_insert';
           notice = json_build_object('queue', NEW.queue, 'state', NEW.state);
 
           -- No point triggering for a job that isn't scheduled to run now
@@ -100,7 +117,7 @@ defmodule Oban.Migrations do
             RETURN null;
           END IF;
         ELSE
-          channel = 'oban_update';
+          channel = '#{prefix}.oban_update';
           notice = json_build_object('queue', NEW.queue, 'new_state', NEW.state, 'old_state', OLD.state);
         END IF;
 
@@ -111,24 +128,24 @@ defmodule Oban.Migrations do
       $$ LANGUAGE plpgsql;
       """
 
-      execute "DROP TRIGGER IF EXISTS oban_notify ON oban_jobs"
+      execute "DROP TRIGGER IF EXISTS oban_notify ON #{prefix}.oban_jobs"
 
       execute """
       CREATE TRIGGER oban_notify
-      AFTER INSERT OR UPDATE ON oban_jobs
-      FOR EACH ROW EXECUTE PROCEDURE oban_jobs_notify();
+      AFTER INSERT OR UPDATE OF state ON #{prefix}.oban_jobs
+      FOR EACH ROW EXECUTE PROCEDURE #{prefix}.oban_jobs_notify();
       """
 
-      execute "COMMENT ON TABLE oban_jobs IS '1'"
+      execute "COMMENT ON TABLE #{prefix}.oban_jobs IS '1'"
     end
 
-    def down do
-      execute "DROP TRIGGER oban_notify ON oban_jobs"
-      execute "DROP FUNCTION oban_jobs_notify()"
+    def down(prefix) do
+      execute "DROP TRIGGER IF EXISTS #{prefix}.oban_notify ON oban_jobs"
+      execute "DROP FUNCTION #{prefix}.oban_jobs_notify()"
 
-      drop_if_exists table("oban_jobs")
+      drop_if_exists table("#{prefix}.oban_jobs")
 
-      execute "DROP TYPE IF EXISTS oban_job_state"
+      execute "DROP TYPE IF EXISTS #{prefix}.oban_job_state"
     end
   end
 
@@ -137,43 +154,48 @@ defmodule Oban.Migrations do
 
     use Ecto.Migration
 
-    def up do
+    def up(prefix) do
       # We only need the scheduled_at index for scheduled and available jobs
-      drop_if_exists index(:oban_jobs, [:scheduled_at])
+      drop_if_exists index(:oban_jobs, [:scheduled_at], prefix: prefix)
+
+      state = "#{prefix}.oban_job_state"
 
       create index(:oban_jobs, [:scheduled_at],
-               where: "state in ('available'::oban_job_state, 'scheduled'::oban_job_state)"
+               where: "state in ('available'::#{state}, 'scheduled'::#{state})",
+               prefix: prefix
              )
 
       create constraint(:oban_jobs, :worker_length,
-               check: "char_length(worker) > 0 AND char_length(worker) < 128"
+               check: "char_length(worker) > 0 AND char_length(worker) < 128",
+               prefix: prefix
              )
 
       create constraint(:oban_jobs, :queue_length,
-               check: "char_length(queue) > 0 AND char_length(queue) < 128"
+               check: "char_length(queue) > 0 AND char_length(queue) < 128",
+               prefix: prefix
              )
 
       execute """
-      CREATE OR REPLACE FUNCTION oban_wrap_id(value bigint) RETURNS int AS $$
+      CREATE OR REPLACE FUNCTION #{prefix}.oban_wrap_id(value bigint) RETURNS int AS $$
       BEGIN
         RETURN (CASE WHEN value > 2147483647 THEN mod(value, 2147483647) ELSE value END)::int;
       END;
       $$ LANGUAGE plpgsql IMMUTABLE;
       """
 
-      execute "COMMENT ON TABLE oban_jobs IS '2'"
+      execute "COMMENT ON TABLE #{prefix}.oban_jobs IS '2'"
     end
 
-    def down do
-      drop_if_exists constraint(:oban_jobs, :queue_length)
-      drop_if_exists constraint(:oban_jobs, :worker_length)
+    def down(prefix) do
+      drop_if_exists constraint(:oban_jobs, :queue_length, prefix: prefix)
+      drop_if_exists constraint(:oban_jobs, :worker_length, prefix: prefix)
 
-      drop_if_exists index(:oban_jobs, [:scheduled_at])
-      create index(:oban_jobs, [:scheduled_at])
+      drop_if_exists index(:oban_jobs, [:scheduled_at], prefix: prefix)
+      create index(:oban_jobs, [:scheduled_at], prefix: prefix)
 
-      execute("DROP FUNCTION oban_wrap_id(value bigint)")
+      execute("DROP FUNCTION #{prefix}.oban_wrap_id(value bigint)")
 
-      execute "COMMENT ON TABLE oban_jobs IS '1'"
+      execute "COMMENT ON TABLE #{prefix}.oban_jobs IS '1'"
     end
   end
 end