+++ test/integration/isolation_test.exs
@@ -7,7 +7,7 @@ defmodule Oban.Integration.IsolationTest do
     start_supervised!({Oban, name: ObanA, repo: Repo, queues: [alpha: 1]}, id: ObanA)
     start_supervised!({Oban, name: ObanB, repo: Repo, queues: [gamma: 1]}, id: ObanB)
 
-    insert_job!(ref: 1, action: "OK")
+    insert_job!(ObanA, ref: 1, action: "OK")
 
     assert_receive {:ok, 1}
 
@@ -15,9 +15,21 @@ defmodule Oban.Integration.IsolationTest do
     stop_supervised(ObanB)
   end
 
-  defp insert_job!(args) do
-    args
-    |> Worker.new(queue: :alpha)
-    |> Repo.insert!()
+  test "inserting and executing jobs with a custom prefix" do
+    start_supervised!({Oban, prefix: "private", repo: Repo, queues: [alpha: 5]})
+
+    job = insert_job!(ref: 1, action: "OK")
+
+    assert Ecto.get_meta(job, :prefix) == "private"
+
+    assert_receive {:ok, 1}
+
+    :ok = stop_supervised(Oban)
+  end
+
+  defp insert_job!(oban \\ Oban, args) do
+    job = Worker.new(args, queue: :alpha)
+
+    Oban.insert!(oban, job)
   end
 end